New eep_stats_dict_dict_count module
Added a new experimental module like eep_stats_dict_count but it also
collects the dicts() that were passed in and not just the counts.
This is really still in alpha as there are a few different design
alternatives that can be tried and this is just one of many options.
So, things will probably change.